theme media

+++ Bump/Data/TestInitializer.cs
@@ -58,7 +58,7 @@ namespace Bump.Data
                     name: $"Theme {i + 1}",
                     content: $"Content of theme {i + 1}",
                     messages: new Message[0],
-                    media: new int[0],
+                    media: new long[0],
                     creationTime: DateTime.Today
                 )
                 {
